third_party/WebKit/LayoutTests/css3/device-adapt/viewport-insert-rule-before-expected.txt

ALERT: viewport size 500x750 scale 0.64 with limits [0.64, 5] and userScalable true
